change minor logic

+++ app/src/main/java/com/rt04/myapplication/presentation/finance/pemasukan/AddIncomeFragment.kt
@@ -81,7 +81,7 @@ class AddIncomeFragment : Fragment(), DatePickerDialog.OnDateSetListener, View.O
         val tanggal = formatDate.format(selectedDate.timeInMillis)
 
         val addIncome = hashMapOf(
-            "jumlah" to nominal,
+            "jumlah" to nominal.toInt(),
             "nama" to nama,
             "tanggal" to tanggal
         )
